Clean up laser control panel
